+++ src/wiremind_kubernetes/kube_config.py
@@ -7,20 +7,7 @@ import kubernetes
 logger = logging.getLogger(__name__)
 
 
-def _get_kube_config_loader_for_yaml_file_with_persistence(filename, persist_config=False, **kwargs):
-
-    kcfg = kubernetes.config.kube_config.KubeConfigMerger(filename)
-    if persist_config and "config_persister" not in kwargs:
-        kwargs["config_persister"] = lambda unused_config_value: kcfg.save_changes()
-
-    return kubernetes.config.kube_config.KubeConfigLoader(config_dict=kcfg.config, config_base_path=None, **kwargs)
-
-
 def _load_kubeconfig():
-    # persist tokens after refreshing them.
-    kubernetes.config.kube_config._get_kube_config_loader_for_yaml_file = (
-        _get_kube_config_loader_for_yaml_file_with_persistence
-    )
     kubernetes.config.load_kube_config()
     logger.debug("Kubernetes configuration successfully set.")